+++ src/azure-cli/HISTORY.rst
@@ -74,6 +74,10 @@ Release History
 * Add new parameters `--enable-delete-retention` and `--delete-retention-days` to support managing delete retention policy for storage account blob-service-properties.
 
 
+**SQL**
+
+* Updating az sql mi update with two new properties tier and family.
+
 2.0.78
 ++++++